+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2025-01-07
+
 ### Added
 - Add support for `classification-results` image pyramids
 
@@ -22,5 +24,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[unreleased]: https://github.com/MaterialsPhysicsANU/anu_ctlab_qemscan_bse_to_zarr3/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/MaterialsPhysicsANU/anu_ctlab_qemscan_bse_to_zarr3/compare/v0.1.1...HEAD
+[0.1.1]: https://github.com/MaterialsPhysicsANU/anu_ctlab_qemscan_bse_to_zarr3/releases/tag/v0.1.1
 [0.1.0]: https://github.com/MaterialsPhysicsANU/anu_ctlab_qemscan_bse_to_zarr3/releases/tag/v0.1.0